pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "golf-analytics"
version = "0.1.0"
description = "Golf performance analytics + Streamlit dashboard"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "streamlit>=1.30",
  "altair>=5.0",
  "pygwalker>=0.4.8",
  "requests>=2.31",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]
api = [
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "supabase>=2.0",
  "python-dotenv>=1.0",
  "python-multipart>=0.0.9",
  "pydantic[email]>=2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/golf_analytics"]

[tool.pytest.ini_options]
testpaths = ["tests"]
